Fix for replacing packages.
Sometimes, we replace a package by another one with a different
name, which causes that the original package is removed.
By removing, we delete all files, that may be owned by the
new package, which is not recognized as a "to be replaced" package
and all files are removed.
So, we now do a diff against the whole file database to check
if an other package owns a file, we want to remove. If so,
we let this file be where it is.