Convert all files in a source tarball to owner root.
Sometimes, when files had crappy owners, there has been
an issue in extrating the properly. To overcome that and
also to overcome that users can take benefit from weird
owners, we will change that to user root.

When extracting the source package, we may replace that
if we feel so.