lexer: Add shell commands with %(echo 123..).
It is possible to use shell commands, just like $(...) in shell.
It's a handy feature when you need some output of a tool
in a variable.

Every time the variable is expanded, the command is called.
It is not ensured that the personality is the same as in
the build environment.