ccache optimizations: Use /tmp and enable compression.
When the cache is on an NFS share, a lot of preprocessed files
will be transferred to that share, because ccache usually uses
$CCACHE_DIR/tmp as directory for its temporary files.
/tmp is always local and only the final data has to be live in
the cache is transferred over the wire.

To decrease the size of the cache, we enable compression. That
will cost a bit of CPU performance, but processors are usually
very fast nowadays and the bottleneck when compiling code is IO.