Perform less parallelism
Sometimes we have trouble with latest versions of g++
that requires too much RAM so that the OOM killer kills
all the compiler processes. Hence we will start less processes
at once for now.

This may to be increased even more in the future as it
seems that g++ requires much more memory with every release.