Replace bogomips by fibonnacci benchmark
BogoMIPS cannot be used as an easy way to determine how fast
a system is any more since there is only dummy values available
in newer kernels.

We use a simple fibonnacci function now which is executed for
a set amount of time (defaults to one second) and count the
number of iterations performed. This is then multiplied by the
number of CPU cores.

Fibonnacci is a very simple test that relies on the integer
unit (no FPU involved) and is not depending on anything else.

This will also allow us to quantitatively determine the
difference in speed between processors.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>