http: Download to a temporary file
This patch makes the HTTP client download data
into memory first and writes it to a temporary
file after 10M of data was downloaded.

This makes downloads faster and reduces IO load
but does not give us any limitation in case bigger
files are downloaded.

We get rid of some file locking which is always
difficult to entertain.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>