Import libpakfire
libpakfire is a C implementation of some functionality
of pakfire. It is supposed to make code cleaner and avoid
too much wrapping in python. It is also intended to make
the code faster where ever performance is critical.

This patch adds a dependency to libarchive which is used
to read and write packages.

This implementation is not fully finished and possibly
rough around the edges. It has to be integrated into
the existing Python code and redundant code will have to
be dropped.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>