libpakfire: parser: Add compatibility for older pakfire files
This patch moves the keyword handling out of the parser and puts
it into the lexer where the "define" keyword is added where it is
needed.

That way, we have a clear definition of a clean syntax without
any shift/reduce or reduce/reduce errors, but remain compatible
to the current pakfire files.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>