libpakfire: db: Set database to WAL mode
We disable autocheckpointing since we are expected to writer rather
large transactions and we would like to keep the database as compact as
possible, too. That is achieved by truncating the WAL log in the end
where the data is being copied into the main database and the WAL log
being emptied after.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>