libpakfire: Drop dependency on libgcrypt
This library was chosen as a lightweight and more easy to use
alternative to OpenSSL.

Unfortunately it does not seem to be up to standard in terms of
reliability, performance and most importantly security.

Therefore we are using OpenSSL which is the de-facto standard library,
well-audited and has a transparent development process.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>